// src/app/directorio/interfaces/anuncio.interface.ts
// Generated by https://quicktype.io

export interface Anuncio {
    id:            number;
    titulo:        string;
    descripcion:   string;
    portada:       string;
    telefono:      string;
    idSeccion:     number;  
    direccion:     string;
    valoracion:    number;
    visitas:       number;
}

export interface ImagenCarrusel{
    image: string;
    thumbImage: string;
    // title: string;
    // alt: string;
}